refactor small tests
