+++ iotcloudsdk/src/androidTest/java/com/kii/iotcloud/IoTCloudAPI_GetTriggerTest.java
@@ -1,0 +1,327 @@
+package com.kii.iotcloud;
+
+import android.support.test.runner.AndroidJUnit4;
+
+import com.kii.iotcloud.command.Action;
+import com.kii.iotcloud.command.Command;
+import com.kii.iotcloud.exception.ForbiddenException;
+import com.kii.iotcloud.exception.NotFoundException;
+import com.kii.iotcloud.exception.ServiceUnavailableException;
+import com.kii.iotcloud.exception.UnsupportedActionException;
+import com.kii.iotcloud.exception.UnsupportedSchemaException;
+import com.kii.iotcloud.internal.InternalUtils;
+import com.kii.iotcloud.schema.Schema;
+import com.kii.iotcloud.schema.SchemaBuilder;
+import com.kii.iotcloud.testschemas.LightState;
+import com.kii.iotcloud.testschemas.SetBrightness;
+import com.kii.iotcloud.testschemas.SetBrightnessResult;
+import com.kii.iotcloud.testschemas.SetColor;
+import com.kii.iotcloud.testschemas.SetColorResult;
+import com.kii.iotcloud.testschemas.SetColorTemperature;
+import com.kii.iotcloud.testschemas.SetColorTemperatureResult;
+import com.kii.iotcloud.testschemas.TurnPower;
+import com.kii.iotcloud.testschemas.TurnPowerResult;
+import com.kii.iotcloud.trigger.Condition;
+import com.kii.iotcloud.trigger.StatePredicate;
+import com.kii.iotcloud.trigger.Trigger;
+import com.kii.iotcloud.trigger.TriggersWhen;
+import com.kii.iotcloud.trigger.clause.Equals;
+import com.squareup.okhttp.mockwebserver.RecordedRequest;
+
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
+
+/**
+ * https://github.com/KiiCorp/IoTCloud/blob/master/rest_api_spec/trigger-endpoint.yaml
+ */
+@RunWith(AndroidJUnit4.class)
+public class IoTCloudAPI_GetTriggerTest extends IoTCloudAPITestBase {
+    @Test
+    public void getTriggerTest() throws Exception {
+        Schema schema = this.createDefaultSchema();
+        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
+        String accessToken = "thing-access-token-1234";
+        String triggerID = "trigger-1234";
+        Target target = new Target(thingID, accessToken);
+
+        List<Action> actions = new ArrayList<Action>();
+        SetColor setColor = new SetColor(128, 0, 255);
+        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
+        actions.add(setColor);
+        actions.add(setColorTemperature);
+        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
+
+        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
+
+        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
+        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, true, "COMMAND_EXECUTION_REJECTED", schema);
+
+        api.setTarget(target);
+        Trigger trigger = api.getTrigger(triggerID);
+        // verify the result
+        Assert.assertEquals(triggerID, trigger.getTriggerID());
+        Assert.assertEquals(true, trigger.disabled());
+        Assert.assertEquals("COMMAND_EXECUTION_REJECTED", trigger.getDisabledReason());
+        this.assertPredicate(predicate, trigger.getPredicate());
+        this.assertCommand(schema, expectedCommand, trigger.getCommand());
+        // verify the request
+        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
+        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
+        Assert.assertEquals("GET", request.getMethod());
+
+        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
+        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
+        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
+        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
+        this.assertRequestHeader(expectedRequestHeaders, request);
+    }
+    @Test
+    public void getTrigger403ErrorTest() throws Exception {
+        Schema schema = this.createDefaultSchema();
+        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
+        String accessToken = "thing-access-token-1234";
+        String triggerID = "trigger-1234";
+        Target target = new Target(thingID, accessToken);
+
+        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
+        this.addEmptyMockResponse(403);
+
+        try {
+            api.setTarget(target);
+            api.getTrigger(triggerID);
+            Assert.fail("IoTCloudRestException should be thrown");
+        } catch (ForbiddenException e) {
+        }
+        // verify the request
+        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
+        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
+        Assert.assertEquals("GET", request.getMethod());
+
+        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
+        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
+        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
+        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
+        this.assertRequestHeader(expectedRequestHeaders, request);
+    }
+    @Test
+    public void getTrigger404ErrorTest() throws Exception {
+        Schema schema = this.createDefaultSchema();
+        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
+        String accessToken = "thing-access-token-1234";
+        String triggerID = "trigger-1234";
+        Target target = new Target(thingID, accessToken);
+
+        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
+        this.addEmptyMockResponse(404);
+
+        try {
+            api.setTarget(target);
+            api.getTrigger(triggerID);
+            Assert.fail("IoTCloudRestException should be thrown");
+        } catch (NotFoundException e) {
+        }
+        // verify the request
+        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
+        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
+        Assert.assertEquals("GET", request.getMethod());
+
+        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
+        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
+        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
+        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
+        this.assertRequestHeader(expectedRequestHeaders, request);
+    }
+    @Test
+    public void getTrigger503ErrorTest() throws Exception {
+        Schema schema = this.createDefaultSchema();
+        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
+        String accessToken = "thing-access-token-1234";
+        String triggerID = "trigger-1234";
+        Target target = new Target(thingID, accessToken);
+
+        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
+        this.addEmptyMockResponse(503);
+
+        try {
+            api.setTarget(target);
+            api.getTrigger(triggerID);
+            Assert.fail("IoTCloudRestException should be thrown");
+        } catch (ServiceUnavailableException e) {
+        }
+        // verify the request
+        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
+        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
+        Assert.assertEquals("GET", request.getMethod());
+
+        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
+        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
+        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
+        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
+        this.assertRequestHeader(expectedRequestHeaders, request);
+    }
+    @Test
+    public void getTriggerWithInvalidSchemaNameTest() throws Exception {
+        Schema schema = this.createDefaultSchema();
+        SchemaBuilder sb = SchemaBuilder.newSchemaBuilder(DEMO_THING_TYPE, DEMO_SCHEMA_NAME + "_invalid", DEMO_SCHEMA_VERSION, LightState.class);
+        sb.addActionClass(SetColor.class, SetColorResult.class);
+        sb.addActionClass(SetBrightness.class, SetBrightnessResult.class);
+        sb.addActionClass(SetColorTemperature.class, SetColorTemperatureResult.class);
+        sb.addActionClass(TurnPower.class, TurnPowerResult.class);
+        Schema invalidSchema = sb.build();
+        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
+        String accessToken = "thing-access-token-1234";
+        String triggerID = "trigger-1234";
+        Target target = new Target(thingID, accessToken);
+
+        List<Action> actions = new ArrayList<Action>();
+        SetColor setColor = new SetColor(128, 0, 255);
+        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
+        actions.add(setColor);
+        actions.add(setColorTemperature);
+        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
+
+        IoTCloudAPI api = this.craeteIoTCloudAPIWithSchema(APP_ID, APP_KEY, invalidSchema);
+
+        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
+        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, true, "COMMAND_EXECUTION_REJECTED", schema);
+
+        try {
+            api.setTarget(target);
+            api.getTrigger(triggerID);
+            Assert.fail("UnsupportedSchemaException should be thrown");
+        } catch (UnsupportedSchemaException e) {
+        }
+        // verify the request
+        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
+        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
+        Assert.assertEquals("GET", request.getMethod());
+
+        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
+        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
+        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
+        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
+        this.assertRequestHeader(expectedRequestHeaders, request);
+    }
+    @Test
+    public void getTriggerWithInvalidSchemaVersionTest() throws Exception {
+        Schema schema = this.createDefaultSchema();
+        SchemaBuilder sb = SchemaBuilder.newSchemaBuilder(DEMO_THING_TYPE, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION + 1, LightState.class);
+        sb.addActionClass(SetColor.class, SetColorResult.class);
+        sb.addActionClass(SetBrightness.class, SetBrightnessResult.class);
+        sb.addActionClass(SetColorTemperature.class, SetColorTemperatureResult.class);
+        sb.addActionClass(TurnPower.class, TurnPowerResult.class);
+        Schema invalidSchema = sb.build();
+        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
+        String accessToken = "thing-access-token-1234";
+        String triggerID = "trigger-1234";
+        Target target = new Target(thingID, accessToken);
+
+        List<Action> actions = new ArrayList<Action>();
+        SetColor setColor = new SetColor(128, 0, 255);
+        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
+        actions.add(setColor);
+        actions.add(setColorTemperature);
+        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
+
+        IoTCloudAPI api = this.craeteIoTCloudAPIWithSchema(APP_ID, APP_KEY, invalidSchema);
+
+        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
+        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, true, "COMMAND_EXECUTION_REJECTED", schema);
+
+        try {
+            api.setTarget(target);
+            api.getTrigger(triggerID);
+            Assert.fail("UnsupportedSchemaException should be thrown");
+        } catch (UnsupportedSchemaException e) {
+        }
+        // verify the request
+        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
+        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
+        Assert.assertEquals("GET", request.getMethod());
+
+        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
+        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
+        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
+        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
+        this.assertRequestHeader(expectedRequestHeaders, request);
+    }
+    @Test
+    public void getTriggerWithUnknownActionTest() throws Exception {
+        Schema schema = this.createDefaultSchema();
+        SchemaBuilder sb = SchemaBuilder.newSchemaBuilder(DEMO_THING_TYPE, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, LightState.class);
+        sb.addActionClass(SetBrightness.class, SetBrightnessResult.class);
+        sb.addActionClass(SetColorTemperature.class, SetColorTemperatureResult.class);
+        sb.addActionClass(TurnPower.class, TurnPowerResult.class);
+        Schema unsupportedSetColorSchema = sb.build();
+        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
+        String accessToken = "thing-access-token-1234";
+        String triggerID = "trigger-1234";
+        Target target = new Target(thingID, accessToken);
+
+        List<Action> actions = new ArrayList<Action>();
+        SetColor setColor = new SetColor(128, 0, 255);
+        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
+        actions.add(setColor);
+        actions.add(setColorTemperature);
+        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
+
+        IoTCloudAPI api = this.craeteIoTCloudAPIWithSchema(APP_ID, APP_KEY, unsupportedSetColorSchema);
+
+        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
+        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, true, "COMMAND_EXECUTION_REJECTED", schema);
+
+        InternalUtils.gsonRepositoryClearCache();
+        try {
+            api.setTarget(target);
+            api.getTrigger(triggerID);
+            Assert.fail("UnsupportedActionException should be thrown");
+        } catch (UnsupportedActionException e) {
+        }
+        // verify the request
+        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
+        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
+        Assert.assertEquals("GET", request.getMethod());
+
+        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
+        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
+        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
+        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
+        this.assertRequestHeader(expectedRequestHeaders, request);
+    }
+    @Test(expected = IllegalStateException.class)
+    public void getTriggerWithNullTargetTest() throws Exception {
+        String triggerID = "trigger-1234";
+
+        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
+        api.getTrigger(triggerID);
+    }
+    @Test(expected = IllegalArgumentException.class)
+    public void getTriggerWithNullTriggerIDTest() throws Exception {
+        Schema schema = this.createDefaultSchema();
+        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
+        String accessToken = "thing-access-token-1234";
+        Target target = new Target(thingID, accessToken);
+
+        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
+        api.setTarget(target);
+        api.getTrigger(null);
+    }
+    @Test(expected = IllegalArgumentException.class)
+    public void getTriggerWithEmptyTriggerIDTest() throws Exception {
+        Schema schema = this.createDefaultSchema();
+        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
+        String accessToken = "thing-access-token-1234";
+        Target target = new Target(thingID, accessToken);
+
+        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
+        api.setTarget(target);
+        api.getTrigger("");
+    }
+}
