--- iotcloudsdk/src/androidTest/java/com/kii/iotcloud/IoTCloudAPI_TriggerTest.java
@@ -1,1496 +1,0 @@
-package com.kii.iotcloud;
-
-import android.support.test.runner.AndroidJUnit4;
-import android.util.Pair;
-
-import com.google.gson.JsonObject;
-import com.kii.iotcloud.command.Action;
-import com.kii.iotcloud.command.Command;
-import com.kii.iotcloud.exception.BadRequestException;
-import com.kii.iotcloud.exception.ForbiddenException;
-import com.kii.iotcloud.exception.NotFoundException;
-import com.kii.iotcloud.exception.ServiceUnavailableException;
-import com.kii.iotcloud.exception.UnsupportedActionException;
-import com.kii.iotcloud.exception.UnsupportedSchemaException;
-import com.kii.iotcloud.internal.GsonRepository;
-import com.kii.iotcloud.internal.InternalUtils;
-import com.kii.iotcloud.schema.Schema;
-import com.kii.iotcloud.schema.SchemaBuilder;
-import com.kii.iotcloud.testschemas.LightState;
-import com.kii.iotcloud.testschemas.SetBrightness;
-import com.kii.iotcloud.testschemas.SetBrightnessResult;
-import com.kii.iotcloud.testschemas.SetColor;
-import com.kii.iotcloud.testschemas.SetColorResult;
-import com.kii.iotcloud.testschemas.SetColorTemperature;
-import com.kii.iotcloud.testschemas.SetColorTemperatureResult;
-import com.kii.iotcloud.testschemas.TurnPower;
-import com.kii.iotcloud.testschemas.TurnPowerResult;
-import com.kii.iotcloud.trigger.Condition;
-import com.kii.iotcloud.trigger.Predicate;
-import com.kii.iotcloud.trigger.StatePredicate;
-import com.kii.iotcloud.trigger.Trigger;
-import com.kii.iotcloud.trigger.TriggersWhen;
-import com.kii.iotcloud.trigger.clause.Equals;
-import com.squareup.okhttp.mockwebserver.RecordedRequest;
-
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.TimeUnit;
-
-/**
- * https://github.com/KiiCorp/IoTCloud/blob/master/rest_api_spec/trigger-endpoint.yaml
- */
-@RunWith(AndroidJUnit4.class)
-public class IoTCloudAPI_TriggerTest extends IoTCloudAPITestBase {
-    @Test
-    public void postNewTriggerTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addMockResponseForPostNewTrigger(201, triggerID);
-        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, false, null, schema);
-
-        api.setTarget(target);
-        Trigger trigger = api.postNewTrigger(DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-        // verify the result
-        Assert.assertEquals(triggerID, trigger.getTriggerID());
-        Assert.assertEquals(false, trigger.disabled());
-        Assert.assertNull(trigger.getDisabledReason());
-        this.assertPredicate(predicate, trigger.getPredicate());
-        this.assertCommand(schema, expectedCommand, trigger.getCommand());
-        // verify the 1st request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers", request1.getPath());
-        Assert.assertEquals("POST", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders1 = new HashMap<String, String>();
-        expectedRequestHeaders1.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders1.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders1.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        expectedRequestHeaders1.put("Content-Type", "application/json");
-        this.assertRequestHeader(expectedRequestHeaders1, request1);
-
-        JsonObject expectedRequestBody = new JsonObject();
-        expectedRequestBody.add("command", GsonRepository.gson(schema).toJsonTree(expectedCommand));
-        expectedRequestBody.add("predicate", GsonRepository.gson(schema).toJsonTree(predicate));
-        this.assertRequestBody(expectedRequestBody, request1);
-
-        // verify the 2nd request
-        RecordedRequest request2 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request2.getPath());
-        Assert.assertEquals("GET", request2.getMethod());
-
-        Map<String, String> expectedRequestHeaders2 = new HashMap<String, String>();
-        expectedRequestHeaders2.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders2.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders2.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders2, request2);
-    }
-    @Test
-    public void postNewTrigger403ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addEmptyMockResponse(403);
-
-        try {
-            api.setTarget(target);
-            api.postNewTrigger(DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ForbiddenException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers", request.getPath());
-        Assert.assertEquals("POST", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders1 = new HashMap<String, String>();
-        expectedRequestHeaders1.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders1.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders1.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        expectedRequestHeaders1.put("Content-Type", "application/json");
-        this.assertRequestHeader(expectedRequestHeaders1, request);
-
-        JsonObject expectedRequestBody = new JsonObject();
-        expectedRequestBody.add("command", GsonRepository.gson(schema).toJsonTree(expectedCommand));
-        expectedRequestBody.add("predicate", GsonRepository.gson(schema).toJsonTree(predicate));
-        this.assertRequestBody(expectedRequestBody, request);
-    }
-    @Test
-    public void postNewTrigger404ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addEmptyMockResponse(404);
-
-        try {
-            api.setTarget(target);
-            api.postNewTrigger(DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (NotFoundException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers", request.getPath());
-        Assert.assertEquals("POST", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders1 = new HashMap<String, String>();
-        expectedRequestHeaders1.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders1.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders1.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        expectedRequestHeaders1.put("Content-Type", "application/json");
-        this.assertRequestHeader(expectedRequestHeaders1, request);
-
-        JsonObject expectedRequestBody = new JsonObject();
-        expectedRequestBody.add("command", GsonRepository.gson(schema).toJsonTree(expectedCommand));
-        expectedRequestBody.add("predicate", GsonRepository.gson(schema).toJsonTree(predicate));
-        this.assertRequestBody(expectedRequestBody, request);
-    }
-    @Test
-    public void postNewTrigger503ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addEmptyMockResponse(503);
-
-        try {
-            api.setTarget(target);
-            api.postNewTrigger(DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ServiceUnavailableException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers", request.getPath());
-        Assert.assertEquals("POST", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders1 = new HashMap<String, String>();
-        expectedRequestHeaders1.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders1.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders1.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        expectedRequestHeaders1.put("Content-Type", "application/json");
-        this.assertRequestHeader(expectedRequestHeaders1, request);
-
-        JsonObject expectedRequestBody = new JsonObject();
-        expectedRequestBody.add("command", GsonRepository.gson(schema).toJsonTree(expectedCommand));
-        expectedRequestBody.add("predicate", GsonRepository.gson(schema).toJsonTree(predicate));
-        this.assertRequestBody(expectedRequestBody, request);
-    }
-    @Test(expected = IllegalStateException.class)
-    public void postNewTriggerWithNullTargetTest() throws Exception {
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.postNewTrigger(DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void postNewTriggerWithNullSchemaNameTest() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.postNewTrigger(null, DEMO_SCHEMA_VERSION, actions, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void postNewTriggerWithEmptySchemaNameTest() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.postNewTrigger("", DEMO_SCHEMA_VERSION, actions, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void postNewTriggerWithNullActionsTest() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.postNewTrigger(DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, null, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void postNewTriggerWithEmptyActionsTest() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.postNewTrigger(DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void postNewTriggerWithNullPredicateTest() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.postNewTrigger(DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, null);
-    }
-    @Test
-    public void getTriggerTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, true, "COMMAND_EXECUTION_REJECTED", schema);
-
-        api.setTarget(target);
-        Trigger trigger = api.getTrigger(triggerID);
-        // verify the result
-        Assert.assertEquals(triggerID, trigger.getTriggerID());
-        Assert.assertEquals(true, trigger.disabled());
-        Assert.assertEquals("COMMAND_EXECUTION_REJECTED", trigger.getDisabledReason());
-        this.assertPredicate(predicate, trigger.getPredicate());
-        this.assertCommand(schema, expectedCommand, trigger.getCommand());
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
-        Assert.assertEquals("GET", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request);
-    }
-    @Test
-    public void getTrigger403ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(403);
-
-        try {
-            api.setTarget(target);
-            api.getTrigger(triggerID);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ForbiddenException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
-        Assert.assertEquals("GET", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request);
-    }
-    @Test
-    public void getTrigger404ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(404);
-
-        try {
-            api.setTarget(target);
-            api.getTrigger(triggerID);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (NotFoundException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
-        Assert.assertEquals("GET", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request);
-    }
-    @Test
-    public void getTrigger503ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(503);
-
-        try {
-            api.setTarget(target);
-            api.getTrigger(triggerID);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ServiceUnavailableException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
-        Assert.assertEquals("GET", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request);
-    }
-    @Test
-    public void getTriggerWithInvalidSchemaNameTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        SchemaBuilder sb = SchemaBuilder.newSchemaBuilder(DEMO_THING_TYPE, DEMO_SCHEMA_NAME + "_invalid", DEMO_SCHEMA_VERSION, LightState.class);
-        sb.addActionClass(SetColor.class, SetColorResult.class);
-        sb.addActionClass(SetBrightness.class, SetBrightnessResult.class);
-        sb.addActionClass(SetColorTemperature.class, SetColorTemperatureResult.class);
-        sb.addActionClass(TurnPower.class, TurnPowerResult.class);
-        Schema invalidSchema = sb.build();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithSchema(APP_ID, APP_KEY, invalidSchema);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, true, "COMMAND_EXECUTION_REJECTED", schema);
-
-        try {
-            api.setTarget(target);
-            api.getTrigger(triggerID);
-            Assert.fail("UnsupportedSchemaException should be thrown");
-        } catch (UnsupportedSchemaException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
-        Assert.assertEquals("GET", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request);
-    }
-    @Test
-    public void getTriggerWithInvalidSchemaVersionTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        SchemaBuilder sb = SchemaBuilder.newSchemaBuilder(DEMO_THING_TYPE, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION + 1, LightState.class);
-        sb.addActionClass(SetColor.class, SetColorResult.class);
-        sb.addActionClass(SetBrightness.class, SetBrightnessResult.class);
-        sb.addActionClass(SetColorTemperature.class, SetColorTemperatureResult.class);
-        sb.addActionClass(TurnPower.class, TurnPowerResult.class);
-        Schema invalidSchema = sb.build();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithSchema(APP_ID, APP_KEY, invalidSchema);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, true, "COMMAND_EXECUTION_REJECTED", schema);
-
-        try {
-            api.setTarget(target);
-            api.getTrigger(triggerID);
-            Assert.fail("UnsupportedSchemaException should be thrown");
-        } catch (UnsupportedSchemaException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
-        Assert.assertEquals("GET", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request);
-    }
-    @Test
-    public void getTriggerWithUnknownActionTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        SchemaBuilder sb = SchemaBuilder.newSchemaBuilder(DEMO_THING_TYPE, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, LightState.class);
-        sb.addActionClass(SetBrightness.class, SetBrightnessResult.class);
-        sb.addActionClass(SetColorTemperature.class, SetColorTemperatureResult.class);
-        sb.addActionClass(TurnPower.class, TurnPowerResult.class);
-        Schema unsupportedSetColorSchema = sb.build();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithSchema(APP_ID, APP_KEY, unsupportedSetColorSchema);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, true, "COMMAND_EXECUTION_REJECTED", schema);
-
-        InternalUtils.gsonRepositoryClearCache();
-        try {
-            api.setTarget(target);
-            api.getTrigger(triggerID);
-            Assert.fail("UnsupportedActionException should be thrown");
-        } catch (UnsupportedActionException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
-        Assert.assertEquals("GET", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request);
-    }
-    @Test(expected = IllegalStateException.class)
-    public void getTriggerWithNullTargetTest() throws Exception {
-        String triggerID = "trigger-1234";
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.getTrigger(triggerID);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void getTriggerWithNullTriggerIDTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.getTrigger(null);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void getTriggerWithEmptyTriggerIDTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.getTrigger("");
-    }
-    @Test
-    public void enableTriggerTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addEmptyMockResponse(204);
-        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, false, null, schema);
-
-        api.setTarget(target);
-        Trigger trigger = api.enableTrigger(triggerID, true);
-        // verify the result
-        Assert.assertEquals(triggerID, trigger.getTriggerID());
-        Assert.assertEquals(false, trigger.disabled());
-        Assert.assertNull(trigger.getDisabledReason());
-        Assert.assertFalse(trigger.disabled());
-        this.assertPredicate(predicate, trigger.getPredicate());
-        this.assertCommand(schema, expectedCommand, trigger.getCommand());
-        // verify the 1st request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID + "/enable", request1.getPath());
-        Assert.assertEquals("PUT", request1.getMethod());
-        // verify the 2nd request
-        RecordedRequest request2 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request2.getPath());
-        Assert.assertEquals("GET", request2.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request2);
-    }
-    @Test
-    public void disableTriggerTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addEmptyMockResponse(204);
-        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, true, null, schema);
-
-        api.setTarget(target);
-        Trigger trigger = api.enableTrigger(triggerID, false);
-        // verify the result
-        Assert.assertEquals(triggerID, trigger.getTriggerID());
-        Assert.assertEquals(true, trigger.disabled());
-        Assert.assertNull(trigger.getDisabledReason());
-        Assert.assertTrue(trigger.disabled());
-        this.assertPredicate(predicate, trigger.getPredicate());
-        this.assertCommand(schema, expectedCommand, trigger.getCommand());
-        // verify the 1st request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID + "/disable", request1.getPath());
-        Assert.assertEquals("PUT", request1.getMethod());
-        // verify the 2nd request
-        RecordedRequest request2 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request2.getPath());
-        Assert.assertEquals("GET", request2.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request2);
-    }
-    @Test
-    public void enableTrigger403Test() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(403);
-        try {
-            api.setTarget(target);
-            api.enableTrigger(triggerID, false);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ForbiddenException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID + "/disable", request.getPath());
-        Assert.assertEquals("PUT", request.getMethod());
-    }
-    @Test
-    public void enableTrigger404Test() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(404);
-        try {
-            api.setTarget(target);
-            api.enableTrigger(triggerID, false);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (NotFoundException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID + "/disable", request.getPath());
-        Assert.assertEquals("PUT", request.getMethod());
-    }
-    @Test
-    public void enableTrigger503Test() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(503);
-        try {
-            api.setTarget(target);
-            api.enableTrigger(triggerID, false);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ServiceUnavailableException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID + "/disable", request.getPath());
-        Assert.assertEquals("PUT", request.getMethod());
-    }
-    @Test(expected = IllegalStateException.class)
-    public void enableTriggerWithNullTargetTest() throws Exception {
-        String triggerID = "trigger-1234";
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.enableTrigger(triggerID, false);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void enableTriggerWithNullTriggerIDTest() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.enableTrigger(null, false);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void enableTriggerWithEmptyTriggerIDTest() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.enableTrigger("", false);
-    }
-    @Test
-    public void listTriggersTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-        String paginationKey = "pagination-12345-key";
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-
-        List<Action> command1Actions = new ArrayList<Action>();
-        command1Actions.add(new TurnPower(true));
-        Command command1 = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), command1Actions);
-        Condition condition1 = new Condition(new Equals("power", true));
-        Predicate predicate1 = new StatePredicate(condition1, TriggersWhen.CONDITION_FALSE_TO_TRUE);
-        Trigger trigger1 = new Trigger(predicate1, command1);
-
-        List<Action> command2Actions = new ArrayList<Action>();
-        command2Actions.add(new SetColor(10, 20, 30));
-        Command command2 = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), command2Actions);
-        Condition condition2 = new Condition(new Equals("power", false));
-        Predicate predicate2 = new StatePredicate(condition2, TriggersWhen.CONDITION_CHANGED);
-        Trigger trigger2 = new Trigger(predicate2, command2);
-
-        List<Action> command3Actions = new ArrayList<Action>();
-        command3Actions.add(new SetColorTemperature(35));
-        command3Actions.add(new SetBrightness(40));
-        Command command3 = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), command3Actions);
-        Condition condition3 = new Condition(new Equals("colorTemperature", 28));
-        Predicate predicate3 = new StatePredicate(condition3, TriggersWhen.CONDITION_TRUE);
-        Trigger trigger3 = new Trigger(predicate3, command3);
-
-        this.addMockResponseForListTriggers(200, new Trigger[] {trigger1, trigger2}, paginationKey, schema);
-        this.addMockResponseForListTriggers(200, new Trigger[]{trigger3}, null, schema);
-
-        // verify the result
-        Pair<List<Trigger>, String> result1 = api.listTriggers(10, null);
-        Assert.assertEquals(paginationKey, result1.second);
-        List<Trigger> triggers1 = result1.first;
-        Assert.assertEquals(2, triggers1.size());
-        this.assertTrigger(schema, trigger1, triggers1.get(0));
-        this.assertTrigger(schema, trigger2, triggers1.get(1));
-
-        Pair<List<Trigger>, String> result2 = api.listTriggers(10, result1.second);
-        Assert.assertNull(result2.second);
-        List<Trigger> triggers2 = result2.first;
-        Assert.assertEquals(1, triggers2.size());
-        this.assertTrigger(schema, trigger3, triggers2.get(0));
-
-        // verify the 1st request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers?bestEffortLimit=10", request1.getPath());
-        Assert.assertEquals("GET", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request1);
-
-        // verify the 2nd request
-        RecordedRequest request2 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers?bestEffortLimit=10&paginationKey=" + paginationKey, request2.getPath());
-        Assert.assertEquals("GET", request2.getMethod());
-        this.assertRequestHeader(expectedRequestHeaders, request2);
-    }
-    @Test
-    public void listTriggersWithBestEffortLimitZeroTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-        String paginationKey = "pagination-12345-key";
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-
-        List<Action> command1Actions = new ArrayList<Action>();
-        command1Actions.add(new TurnPower(true));
-        Command command1 = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), command1Actions);
-        Condition condition1 = new Condition(new Equals("power", true));
-        Predicate predicate1 = new StatePredicate(condition1, TriggersWhen.CONDITION_FALSE_TO_TRUE);
-        Trigger trigger1 = new Trigger(predicate1, command1);
-
-        List<Action> command2Actions = new ArrayList<Action>();
-        command2Actions.add(new SetColor(10, 20, 30));
-        Command command2 = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), command2Actions);
-        Condition condition2 = new Condition(new Equals("power", false));
-        Predicate predicate2 = new StatePredicate(condition2, TriggersWhen.CONDITION_CHANGED);
-        Trigger trigger2 = new Trigger(predicate2, command2);
-
-        List<Action> command3Actions = new ArrayList<Action>();
-        command3Actions.add(new SetColorTemperature(35));
-        command3Actions.add(new SetBrightness(40));
-        Command command3 = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), command3Actions);
-        Condition condition3 = new Condition(new Equals("colorTemperature", 28));
-        Predicate predicate3 = new StatePredicate(condition3, TriggersWhen.CONDITION_TRUE);
-        Trigger trigger3 = new Trigger(predicate3, command3);
-
-        this.addMockResponseForListTriggers(200, new Trigger[] {trigger1, trigger2}, paginationKey, schema);
-        this.addMockResponseForListTriggers(200, new Trigger[]{trigger3}, null, schema);
-
-        // verify the result
-        Pair<List<Trigger>, String> result1 = api.listTriggers(0, null);
-        Assert.assertEquals(paginationKey, result1.second);
-        List<Trigger> triggers1 = result1.first;
-        Assert.assertEquals(2, triggers1.size());
-        this.assertTrigger(schema, trigger1, triggers1.get(0));
-        this.assertTrigger(schema, trigger2, triggers1.get(1));
-
-        Pair<List<Trigger>, String> result2 = api.listTriggers(0, result1.second);
-        Assert.assertNull(result2.second);
-        List<Trigger> triggers2 = result2.first;
-        Assert.assertEquals(1, triggers2.size());
-        this.assertTrigger(schema, trigger3, triggers2.get(0));
-
-        // verify the 1st request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers", request1.getPath());
-        Assert.assertEquals("GET", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request1);
-
-        // verify the 2nd request
-        RecordedRequest request2 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers?paginationKey=" + paginationKey, request2.getPath());
-        Assert.assertEquals("GET", request2.getMethod());
-        this.assertRequestHeader(expectedRequestHeaders, request2);
-    }
-    @Test
-    public void listTriggers400ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(400);
-        try {
-            api.setTarget(target);
-            api.listTriggers(10, null);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (BadRequestException e) {
-        }
-        // verify the request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers?bestEffortLimit=10", request1.getPath());
-        Assert.assertEquals("GET", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request1);
-    }
-    @Test
-    public void listTriggers403ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(403);
-        try {
-            api.setTarget(target);
-            api.listTriggers(10, null);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ForbiddenException e) {
-        }
-        // verify the request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers?bestEffortLimit=10", request1.getPath());
-        Assert.assertEquals("GET", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request1);
-    }
-    @Test
-    public void listTriggers404ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(404);
-        try {
-            api.setTarget(target);
-            api.listTriggers(10, null);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (NotFoundException e) {
-        }
-        // verify the request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers?bestEffortLimit=10", request1.getPath());
-        Assert.assertEquals("GET", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request1);
-    }
-    @Test
-    public void listTriggers503ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(503);
-        try {
-            api.setTarget(target);
-            api.listTriggers(10, null);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ServiceUnavailableException e) {
-        }
-        // verify the request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers?bestEffortLimit=10", request1.getPath());
-        Assert.assertEquals("GET", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request1);
-    }
-    @Test(expected = IllegalStateException.class)
-    public void listTriggersWithNullTargetTest() throws Exception {
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.listTriggers(10, null);
-    }
-    @Test
-    public void deleteTriggerTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, false, null, schema);
-        this.addEmptyMockResponse(204);
-
-        api.setTarget(target);
-        Trigger trigger = api.deleteTrigger(triggerID);
-        // verify the result
-        Assert.assertEquals(triggerID, trigger.getTriggerID());
-        Assert.assertEquals(false, trigger.disabled());
-        Assert.assertNull(trigger.getDisabledReason());
-        Assert.assertFalse(trigger.disabled());
-        this.assertPredicate(predicate, trigger.getPredicate());
-        this.assertCommand(schema, expectedCommand, trigger.getCommand());
-        // verify the 1st request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request1.getPath());
-        Assert.assertEquals("GET", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request1);
-        // verify the 2nd request
-        RecordedRequest request2 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request2.getPath());
-        Assert.assertEquals("DELETE", request2.getMethod());
-        this.assertRequestHeader(expectedRequestHeaders, request2);
-    }
-    @Test
-    public void deleteTrigger403ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(403);
-
-        try {
-            api.setTarget(target);
-            api.deleteTrigger(triggerID);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ForbiddenException e) {
-        }
-        // verify the 1st request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request1.getPath());
-        Assert.assertEquals("GET", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request1);
-    }
-    @Test
-    public void deleteTrigger404ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(404);
-
-        try {
-            api.setTarget(target);
-            api.deleteTrigger(triggerID);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (NotFoundException e) {
-        }
-        // verify the 1st request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request1.getPath());
-        Assert.assertEquals("GET", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request1);
-    }
-    @Test
-    public void deleteTrigger503ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        this.addEmptyMockResponse(503);
-
-        try {
-            api.setTarget(target);
-            api.deleteTrigger(triggerID);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ServiceUnavailableException e) {
-        }
-        // verify the 1st request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request1.getPath());
-        Assert.assertEquals("GET", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders, request1);
-    }
-    @Test(expected = IllegalStateException.class)
-    public void deleteTriggerWithNullTargetTest() throws Exception {
-        String triggerID = "trigger-1234";
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.deleteTrigger(triggerID);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void deleteTriggerWithNullTriggerIDTest() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.deleteTrigger(null);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void deleteTriggerWithEmptyTriggerIDTest() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.deleteTrigger("");
-    }
-    @Test
-    public void patchTriggerTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addEmptyMockResponse(204);
-        this.addMockResponseForGetTrigger(200, triggerID, expectedCommand, predicate, false, null, schema);
-
-        api.setTarget(target);
-        Trigger trigger = api.patchTrigger(triggerID, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-        // verify the result
-        Assert.assertEquals(triggerID, trigger.getTriggerID());
-        Assert.assertEquals(false, trigger.disabled());
-        Assert.assertNull(trigger.getDisabledReason());
-        this.assertPredicate(predicate, trigger.getPredicate());
-        this.assertCommand(schema, expectedCommand, trigger.getCommand());
-        // verify the 1st request
-        RecordedRequest request1 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request1.getPath());
-        Assert.assertEquals("PATCH", request1.getMethod());
-
-        Map<String, String> expectedRequestHeaders1 = new HashMap<String, String>();
-        expectedRequestHeaders1.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders1.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders1.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        expectedRequestHeaders1.put("Content-Type", "application/json");
-        this.assertRequestHeader(expectedRequestHeaders1, request1);
-
-        JsonObject expectedRequestBody = new JsonObject();
-        expectedRequestBody.add("command", GsonRepository.gson(schema).toJsonTree(expectedCommand));
-        expectedRequestBody.add("predicate", GsonRepository.gson(schema).toJsonTree(predicate));
-        this.assertRequestBody(expectedRequestBody, request1);
-
-        // verify the 2nd request
-        RecordedRequest request2 = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request2.getPath());
-        Assert.assertEquals("GET", request2.getMethod());
-
-        Map<String, String> expectedRequestHeaders2 = new HashMap<String, String>();
-        expectedRequestHeaders2.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders2.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders2.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        this.assertRequestHeader(expectedRequestHeaders2, request2);
-    }
-    @Test
-    public void patchTrigger403ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addEmptyMockResponse(403);
-
-        try {
-            api.setTarget(target);
-            api.patchTrigger(triggerID, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ForbiddenException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
-        Assert.assertEquals("PATCH", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        expectedRequestHeaders.put("Content-Type", "application/json");
-        this.assertRequestHeader(expectedRequestHeaders, request);
-
-        JsonObject expectedRequestBody = new JsonObject();
-        expectedRequestBody.add("command", GsonRepository.gson(schema).toJsonTree(expectedCommand));
-        expectedRequestBody.add("predicate", GsonRepository.gson(schema).toJsonTree(predicate));
-        this.assertRequestBody(expectedRequestBody, request);
-    }
-    @Test
-    public void patchTrigger404ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addEmptyMockResponse(404);
-
-        try {
-            api.setTarget(target);
-            api.patchTrigger(triggerID, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (NotFoundException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
-        Assert.assertEquals("PATCH", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        expectedRequestHeaders.put("Content-Type", "application/json");
-        this.assertRequestHeader(expectedRequestHeaders, request);
-
-        JsonObject expectedRequestBody = new JsonObject();
-        expectedRequestBody.add("command", GsonRepository.gson(schema).toJsonTree(expectedCommand));
-        expectedRequestBody.add("predicate", GsonRepository.gson(schema).toJsonTree(predicate));
-        this.assertRequestBody(expectedRequestBody, request);
-    }
-    @Test
-    public void patchTrigger503ErrorTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-
-        Command expectedCommand = new Command(schema.getSchemaName(), schema.getSchemaVersion(), target.getID(), api.getOwner().getID(), actions);
-        this.addEmptyMockResponse(503);
-
-        try {
-            api.setTarget(target);
-            api.patchTrigger(triggerID, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-            Assert.fail("IoTCloudRestException should be thrown");
-        } catch (ServiceUnavailableException e) {
-        }
-        // verify the request
-        RecordedRequest request = this.server.takeRequest(1, TimeUnit.SECONDS);
-        Assert.assertEquals(BASE_PATH + "/targets/" + thingID.toString() + "/triggers/" + triggerID, request.getPath());
-        Assert.assertEquals("PATCH", request.getMethod());
-
-        Map<String, String> expectedRequestHeaders = new HashMap<String, String>();
-        expectedRequestHeaders.put("X-Kii-AppID", APP_ID);
-        expectedRequestHeaders.put("X-Kii-AppKey", APP_KEY);
-        expectedRequestHeaders.put("Authorization", "Bearer " + api.getOwner().getAccessToken());
-        expectedRequestHeaders.put("Content-Type", "application/json");
-        this.assertRequestHeader(expectedRequestHeaders, request);
-
-        JsonObject expectedRequestBody = new JsonObject();
-        expectedRequestBody.add("command", GsonRepository.gson(schema).toJsonTree(expectedCommand));
-        expectedRequestBody.add("predicate", GsonRepository.gson(schema).toJsonTree(predicate));
-        this.assertRequestBody(expectedRequestBody, request);
-    }
-    @Test(expected = IllegalStateException.class)
-    public void patchTriggerWithNullTargetTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        String triggerID = "trigger-1234";
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.patchTrigger(triggerID, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void patchTriggerWithNullTriggerIDTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.patchTrigger(null, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void patchTriggerWithEmptyTriggerIDTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.patchTrigger(null, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void patchTriggerWithNullSchemaNameTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.patchTrigger(triggerID, null, DEMO_SCHEMA_VERSION, actions, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void patchTriggerWithEmptySchemaNameTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.patchTrigger(triggerID, "", DEMO_SCHEMA_VERSION, actions, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void patchTriggerWithNullActionsTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.patchTrigger(triggerID, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, null, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void patchTriggerWithEmptyActionsTest() throws Exception {
-        Schema schema = this.createDefaultSchema();
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        StatePredicate predicate = new StatePredicate(new Condition(new Equals("power", true)), TriggersWhen.CONDITION_CHANGED);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.patchTrigger(triggerID, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, predicate);
-    }
-    @Test(expected = IllegalArgumentException.class)
-    public void patchTriggerWithNullPredicateTest() throws Exception {
-        TypedID thingID = new TypedID(TypedID.Types.THING, "th.1234567890");
-        String accessToken = "thing-access-token-1234";
-        String triggerID = "trigger-1234";
-        Target target = new Target(thingID, accessToken);
-
-        List<Action> actions = new ArrayList<Action>();
-        SetColor setColor = new SetColor(128, 0, 255);
-        SetColorTemperature setColorTemperature = new SetColorTemperature(25);
-        actions.add(setColor);
-        actions.add(setColorTemperature);
-
-        IoTCloudAPI api = this.craeteIoTCloudAPIWithDemoSchema(APP_ID, APP_KEY);
-        api.setTarget(target);
-        api.patchTrigger(triggerID, DEMO_SCHEMA_NAME, DEMO_SCHEMA_VERSION, actions, null);
-    }
-}
-
-
